requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
typing
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pymongo
jinja2
boto3
pandas